+++ tower-reqwest/src/adapters/mod.rs
@@ -1,5 +1,3 @@
 //! Adapters for various clients
 
 pub mod reqwest;
-#[cfg(feature = "reqwest-middleware")]
-mod reqwest_middleware;